Make repository private?
Hello,

I am writing to ask if you might be willing to make this repository private.

I'm an instructor for this course. I noticed that this project is publicly available. It contains a solution to Project 2. We think Project 2 is so valuable that we'd like to be able to offer it again in future semesters, but having solutions (source code and design docs) publicly available creates a temptation to future students. I'm dealing with some academic misconduct cases right now as a result of this sort of thing, and I hate to see students go through that.

Would you be willing to preserve the value of this project for future students, by making your repository private, so that your solution is not publicly available? I think this would help future students learn more and avoid the pain of students getting tempted to cheat, and then getting caught and going through pain and suffering as a result.

Thank you for considering it. Happy new year!

Regards,

-- David Wagner